refactor(NamingConvention)[NI]: users -> userz
